refactor(trpc): extract admin procedures into routers/admin
Continues p2-22 by carving the admin-only procedures out of
`src/trpc/index.ts` into `src/trpc/routers/admin.ts`:

- updateUserTier
- deleteUser
- getSystemMetrics
- getErrorLogs

Same flat-API-preserving pattern as the highlight extraction: the new
module exports a plain `adminProcedures` object that the root router
spreads in. No public API changes, no behavior changes.

Also drops the now-unused `adminProcedure` import from index.ts so the
file no longer holds onto procedures it doesn't define.

Verification: `admin`, `index`, `ownership`, `sharing`, and `billing`
tRPC tests all pass (28/28). The existing logger.warn output during
admin tests is the procedure's audit log, not a failure.

Refs: p2-22-split-trpc
